✨ Creación de carpeta ViewModel y recolocación de estos en dicha carpeta.
