+++ CentroDeSalud/Models/ViewModels/ErrorViewModel.cs
@@ -1,8 +1,8 @@
-namespace CentroDeSalud.Models
+namespace CentroDeSalud.Models.ViewModels
 {
     public class ErrorViewModel
     {
-        public string? RequestId { get; set; }
+        public string RequestId { get; set; }
 
         public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
     }
