+++ CentroDeSalud/Views/Shared/Error.cshtml
@@ -1,4 +1,5 @@
-﻿@model ErrorViewModel
+﻿@using CentroDeSalud.Models.ViewModels
+@model ErrorViewModel
 @{
     ViewData["Title"] = "Error";
 }
